app/src/main/java/io/clfm/multitimer/data/TimerDatabase.kt
package io.clfm.multitimer.data

import android.content.Context
import androidx.room.Database
import androidx.room.Room
import androidx.room.RoomDatabase
import androidx.room.TypeConverters
import io.clfm.multitimer.data.converter.InstantConverter
import io.clfm.multitimer.data.converter.TimerStateConverter

@Database(entities = [Timer::class], version = 1, exportSchema = false)
@TypeConverters(InstantConverter::class, TimerStateConverter::class)
abstract class TimerDatabase : RoomDatabase() {

    abstract fun timerDao(): TimerDao

    companion object {
        @Volatile
        private var INSTANCE: TimerDatabase? = null

        fun getDatabase(context: Context): TimerDatabase {
            return INSTANCE ?: synchronized(this) {
                val instance = Room.databaseBuilder(
                    context.applicationContext,
                    TimerDatabase::class.java,
                    "timer_database"
                )
                    .fallbackToDestructiveMigration()
                    .build()
                INSTANCE = instance
                instance
            }
        }
    }

}
